+++ CHANGES.md
@@ -2,13 +2,15 @@
 
 ## v0.1.0
 
-- Fix ``scalyr.query`` action so it works correctly. Also fix parameter name from 
+- Fix ``scalyr.query`` action so it works correctly. Also fix parameter name from
   ``fromTime`` / ``toTime`` to ``startTime`` / ``endTime`` and add new ``priority``
   and ``columns`` parameter.
 - Add new ``scalyr.power_query`` action.
 - Add new ``scalyr.timeseries_query`` action.
 - Add new ``scalyr.timeseries_queries`` action.
 - Add new action aliases.
+- In addition to ``token`` also allow base API endpoint url to be overriden on action
+  execution basis using ``api_url`` action parameter.
 
 ## v0.0.1